feat(edit-content) fix sq issue #30022
