+++ core-web/libs/edit-content/src/lib/utils/mocks.ts
@@ -13,7 +13,6 @@ import {
     DotCMSContentTypeField,
     DotCMSContentTypeLayoutRow,
     DotCMSTempFile,
-    DotCMSWorkflowAction,
     FeaturedFlags
 } from '@dotcms/dotcms-models';
 import { MockDotMessageService } from '@dotcms/utils-testing';
@@ -35,11 +34,11 @@ export const TEXT_FIELD_MOCK: DotCMSContentTypeField = {
     fieldTypeLabel: 'Text',
     fieldVariables: [],
     fixed: false,
+    hint: 'A helper text',
     iDate: 1696896882000,
     id: 'c3b928bc2b59fc22c67022de4dd4b5c4',
     indexed: false,
     listed: false,
-    hint: 'A helper text',
     modDate: 1696896882000,
     name: 'testVariable',
     readOnly: false,
@@ -84,8 +83,8 @@ export const SELECT_FIELD_TEXT_MOCK = {
     fieldVariables: [],
     fixed: false,
     forceIncludeInApi: false,
-    iDate: 1697579843000,
     hint: 'A hint Text',
+    iDate: 1697579843000,
     id: 'a6f33b8941b6c06c8ab36e44c4bf6500',
     indexed: false,
     listed: false,
@@ -108,8 +107,8 @@ export const SELECT_FIELD_BOOLEAN_MOCK = {
     fieldTypeLabel: 'Select',
     fieldVariables: [],
     fixed: false,
-    hint: 'A hint Text',
     forceIncludeInApi: false,
+    hint: 'A hint Text',
     iDate: 1697661273000,
     id: '8c5648fe4dedc06baf314f362c00431b',
     indexed: false,
@@ -131,10 +130,10 @@ export const SELECT_FIELD_FLOAT_MOCK = {
     dataType: 'FLOAT',
     fieldType: 'Select',
     fieldTypeLabel: 'Select',
-    hint: 'A hint Text',
     fieldVariables: [],
     fixed: false,
     forceIncludeInApi: false,
+    hint: 'A hint Text',
     iDate: 1697661848000,
     id: '8c2edc3ee461fa50041a9e5831f1a86a',
     indexed: false,
@@ -159,6 +158,7 @@ export const SELECT_FIELD_INTEGER_MOCK = {
     fieldVariables: [],
     fixed: false,
     forceIncludeInApi: false,
+    hint: 'A hint Text',
     iDate: 1697662296000,
     id: '89bdd8e525ef9a4c923f4b54d9a0e4f8',
     indexed: false,
@@ -168,7 +168,6 @@ export const SELECT_FIELD_INTEGER_MOCK = {
     readOnly: false,
     required: false,
     searchable: false,
-    hint: 'A hint Text',
     sortOrder: 6,
     unique: false,
     values: 'One hundred|100\r\nOne thousand|1000\r\nTen thousand|10000',
@@ -184,6 +183,7 @@ export const RADIO_FIELD_TEXT_MOCK = {
     fieldVariables: [],
     fixed: false,
     forceIncludeInApi: false,
+    hint: 'A hint Text',
     iDate: 1697598313000,
     id: '824b4e9907fe4f450ced438598cc0ce8',
     indexed: false,
@@ -192,7 +192,6 @@ export const RADIO_FIELD_TEXT_MOCK = {
     name: 'radio',
     readOnly: false,
     required: false,
-    hint: 'A hint Text',
     searchable: false,
     sortOrder: 8,
     unique: false,
@@ -209,6 +208,7 @@ export const RADIO_FIELD_BOOLEAN_MOCK = {
     fieldVariables: [],
     fixed: false,
     forceIncludeInApi: false,
+    hint: 'A hint Text',
     iDate: 1697656862000,
     id: 'e4b3ef6a8cb50ff77fe2534c2b237d71',
     indexed: false,
@@ -219,7 +219,6 @@ export const RADIO_FIELD_BOOLEAN_MOCK = {
     required: false,
     searchable: false,
     sortOrder: 9,
-    hint: 'A hint Text',
     unique: false,
     values: 'Falsy|false\r\nTruthy|true',
     variable: 'radioTrueFalse'
@@ -235,6 +234,7 @@ export const RADIO_FIELD_FLOAT_MOCK = {
     fieldVariables: [],
     fixed: false,
     forceIncludeInApi: false,
+    hint: 'A hint Text',
     iDate: 1697656895000,
     id: 'b26138321e5a449cdf7b73f927643016',
     indexed: false,
@@ -242,7 +242,6 @@ export const RADIO_FIELD_FLOAT_MOCK = {
     modDate: 1697662296000,
     name: 'radioDecimal',
     readOnly: false,
-    hint: 'A hint Text',
     required: false,
     searchable: false,
     sortOrder: 10,
@@ -261,6 +260,7 @@ export const RADIO_FIELD_INTEGER_MOCK = {
     fieldVariables: [],
     fixed: false,
     forceIncludeInApi: false,
+    hint: 'A hint Text',
     iDate: 1697656956000,
     id: 'bdd0f00375e23b7a64608d78c8fcb2dc',
     indexed: false,
@@ -268,7 +268,6 @@ export const RADIO_FIELD_INTEGER_MOCK = {
     modDate: 1697662296000,
     name: 'radioWholeNumber',
     readOnly: false,
-    hint: 'A hint Text',
     required: true,
     searchable: false,
     sortOrder: 11,
@@ -285,10 +284,10 @@ export const DATE_FIELD_MOCK: DotCMSContentTypeField = {
     fieldTypeLabel: 'Date',
     fieldVariables: [],
     fixed: false,
+    hint: 'A hint text',
     iDate: 1698250833000,
     id: '55b4fdcf51eddf01b0f462384e8b3439',
     indexed: false,
-    hint: 'A hint text',
     listed: false,
     modDate: 1698250833000,
     name: 'date',
@@ -309,11 +308,11 @@ export const DATE_AND_TIME_FIELD_MOCK: DotCMSContentTypeField = {
     fieldTypeLabel: 'Date and Time',
     fieldVariables: [],
     fixed: false,
+    hint: 'A hint text',
     iDate: 1698250840000,
     id: '9e669bacc84ce6530bba5f295becc76c',
     indexed: false,
     listed: false,
-    hint: 'A hint text',
     modDate: 1698250840000,
     name: 'date and time',
     readOnly: false,
@@ -325,6 +324,7 @@ export const DATE_AND_TIME_FIELD_MOCK: DotCMSContentTypeField = {
 };
 
 export const TIME_FIELD_MOCK: DotCMSContentTypeField = {
+    id: '1005cde03b962dd0ce7bb4c4ec97f89c',
     clazz: 'com.dotcms.contenttype.model.field.ImmutableTimeField',
     contentTypeId: '4d22214338844b4aed0367933e9bf500',
     dataType: 'DATE',
@@ -334,7 +334,6 @@ export const TIME_FIELD_MOCK: DotCMSContentTypeField = {
     fixed: false,
     hint: 'A hint text',
     iDate: 1698250847000,
-    id: '1005cde03b962dd0ce7bb4c4ec97f89c',
     indexed: false,
     listed: false,
     modDate: 1698250847000,
@@ -348,6 +347,7 @@ export const TIME_FIELD_MOCK: DotCMSContentTypeField = {
 };
 
 export const TAG_FIELD_MOCK: DotCMSContentTypeField = {
+    id: '1ba4927b83aae5b17921679053b0b5fe',
     clazz: 'com.dotcms.contenttype.model.field.ImmutableTagField',
     contentTypeId: '61226fd915b7f025da020fc1f5856ab7',
     dataType: 'SYSTEM',
@@ -358,7 +358,6 @@ export const TAG_FIELD_MOCK: DotCMSContentTypeField = {
     fixed: false,
     hint: 'Some hint',
     iDate: 1698346136000,
-    id: '1ba4927b83aae5b17921679053b0b5fe',
     indexed: true,
     listed: false,
     modDate: 1698346136000,
@@ -372,6 +371,7 @@ export const TAG_FIELD_MOCK: DotCMSContentTypeField = {
 };
 
 export const CHECKBOX_FIELD_MOCK: DotCMSContentTypeField = {
+    id: '96909fa20a00497ce3b766b52edac0ec',
     clazz: 'com.dotcms.contenttype.model.field.ImmutableCheckboxField',
     contentTypeId: '93ebaff75f3e3887bea73eca04588dc9',
     dataType: 'TEXT',
@@ -381,7 +381,6 @@ export const CHECKBOX_FIELD_MOCK: DotCMSContentTypeField = {
     fixed: false,
     hint: 'A hint text',
     iDate: 1698291913000,
-    id: '96909fa20a00497ce3b766b52edac0ec',
     indexed: false,
     listed: false,
     modDate: 1698291913000,
@@ -396,6 +395,7 @@ export const CHECKBOX_FIELD_MOCK: DotCMSContentTypeField = {
 };
 
 export const MULTI_SELECT_FIELD_MOCK: DotCMSContentTypeField = {
+    id: '535a6de288e3fe91fad2679e8d7d966b',
     clazz: 'com.dotcms.contenttype.model.field.ImmutableMultiSelectField',
     contentTypeId: '93ebaff75f3e3887bea73eca04588dc9',
     dataType: 'LONG_TEXT',
@@ -405,7 +405,6 @@ export const MULTI_SELECT_FIELD_MOCK: DotCMSContentTypeField = {
     fixed: false,
     hint: 'A hint text',
     iDate: 1698264695000,
-    id: '535a6de288e3fe91fad2679e8d7d966b',
     indexed: false,
     listed: false,
     modDate: 1698291913000,
@@ -420,6 +419,7 @@ export const MULTI_SELECT_FIELD_MOCK: DotCMSContentTypeField = {
 };
 
 export const BLOCK_EDITOR_FIELD_MOCK: DotCMSContentTypeField = {
+    id: '71fe962eb681c5ffd6cd1623e5fc575a',
     clazz: 'com.dotcms.contenttype.model.field.ImmutableStoryBlockField',
     contentTypeId: '799f176a-d32e-4844-a07c-1b5fcd107578',
     dataType: 'LONG_TEXT',
@@ -436,7 +436,6 @@ export const BLOCK_EDITOR_FIELD_MOCK: DotCMSContentTypeField = {
     ],
     fixed: false,
     iDate: 1649791703000,
-    id: '71fe962eb681c5ffd6cd1623e5fc575a',
     indexed: false,
     listed: false,
     hint: 'A helper text',
@@ -451,6 +450,7 @@ export const BLOCK_EDITOR_FIELD_MOCK: DotCMSContentTypeField = {
 };
 
 export const BINARY_FIELD_MOCK: DotCMSContentTypeField = {
+    id: '5df3f8fc49177c195740bcdc02ec2db7',
     clazz: 'com.dotcms.contenttype.model.field.ImmutableBinaryField',
     contentTypeId: 'd1901a41d38b6686dd5ed8f910346d7a',
     dataType: 'SYSTEM',
@@ -476,7 +476,6 @@ export const BINARY_FIELD_MOCK: DotCMSContentTypeField = {
     fixed: false,
     forceIncludeInApi: false,
     iDate: 1698153564000,
-    id: '5df3f8fc49177c195740bcdc02ec2db7',
     indexed: false,
     listed: false,
     modDate: 1698153564000,
@@ -490,6 +489,7 @@ export const BINARY_FIELD_MOCK: DotCMSContentTypeField = {
 };
 
 export const IMAGE_FIELD_MOCK: DotCMSContentTypeField = {
+    id: 'fec3e11696cf9b0f99139c160a598e02',
     clazz: 'com.dotcms.contenttype.model.field.ImmutableImageField',
     contentTypeId: 'a8f941d835e4b4f3e4e71b45add34c60',
     dataType: 'TEXT',
@@ -499,7 +499,6 @@ export const IMAGE_FIELD_MOCK: DotCMSContentTypeField = {
     fixed: false,
     forceIncludeInApi: false,
     iDate: 1726517012000,
-    id: 'fec3e11696cf9b0f99139c160a598e02',
     indexed: false,
     listed: false,
     modDate: 1726517012000,
@@ -514,6 +513,7 @@ export const IMAGE_FIELD_MOCK: DotCMSContentTypeField = {
 };
 
 export const FILE_FIELD_MOCK: DotCMSContentTypeField = {
+    id: 'f90afb1384e04507ba03e8701f7e4000',
     clazz: 'com.dotcms.contenttype.model.field.ImmutableFileField',
     contentTypeId: 'a8f941d835e4b4f3e4e71b45add34c60',
     dataType: 'TEXT',
@@ -523,7 +523,6 @@ export const FILE_FIELD_MOCK: DotCMSContentTypeField = {
     fixed: false,
     forceIncludeInApi: false,
     iDate: 1726507692000,
-    id: 'f90afb1384e04507ba03e8701f7e4000',
     indexed: false,
     listed: false,
     modDate: 1726517016000,
@@ -538,6 +537,7 @@ export const FILE_FIELD_MOCK: DotCMSContentTypeField = {
 };
 
 export const CUSTOM_FIELD_MOCK: DotCMSContentTypeField = {
+    id: '64d5c84f04df900c79a94e087c6fed05',
     clazz: 'com.dotcms.contenttype.model.field.ImmutableCustomField',
     contentTypeId: '61226fd915b7f025da020fc1f5856ab7',
     dataType: 'LONG_TEXT',
@@ -546,7 +546,6 @@ export const CUSTOM_FIELD_MOCK: DotCMSContentTypeField = {
     fieldVariables: [],
     fixed: false,
     iDate: 1700516848000,
-    id: '64d5c84f04df900c79a94e087c6fed05',
     indexed: false,
     listed: false,
     modDate: 1700622670000,
@@ -562,6 +561,7 @@ export const CUSTOM_FIELD_MOCK: DotCMSContentTypeField = {
 };
 
 export const JSON_FIELD_MOCK: DotCMSContentTypeField = {
+    id: '96909fa20a00497cd3b766b52edac0ec',
     clazz: 'com.dotcms.contenttype.model.field.ImmutableJSONField',
     contentTypeId: '93ebaff75f3e3887bea73ecd04588dc9',
     dataType: 'TEXT',
@@ -571,7 +571,6 @@ export const JSON_FIELD_MOCK: DotCMSContentTypeField = {
     fixed: false,
     hint: 'A hint text',
     iDate: 1698291913000,
-    id: '96909fa20a00497cd3b766b52edac0ec',
     indexed: false,
     listed: false,
     modDate: 1698291913000,
@@ -586,6 +585,7 @@ export const JSON_FIELD_MOCK: DotCMSContentTypeField = {
 };
 
 export const KEY_VALUE_MOCK: DotCMSContentTypeField = {
+    id: '96909fa20a00497cd3b766b52edac0ec',
     clazz: 'com.dotcms.contenttype.model.field.ImmutableJSONField',
     contentTypeId: '93ebaff75f3e3887bea73ecd04588dc9',
     dataType: 'TEXT',
@@ -595,7 +595,6 @@ export const KEY_VALUE_MOCK: DotCMSContentTypeField = {
     fixed: false,
     hint: 'A hint text',
     iDate: 1698291913000,
-    id: '96909fa20a00497cd3b766b52edac0ec',
     indexed: false,
     listed: false,
     modDate: 1698291913000,
@@ -610,6 +609,7 @@ export const KEY_VALUE_MOCK: DotCMSContentTypeField = {
 };
 
 export const HOST_FOLDER_TEXT_MOCK = {
+    id: 'b7c41ffd6b6bc1250f2fc85a3637471b',
     clazz: 'com.dotcms.contenttype.model.field.ImmutableHostFolderField',
     contentTypeId: '61226fd915b7f025da020fc1f5856ab7',
     dataType: 'SYSTEM',
@@ -620,7 +620,6 @@ export const HOST_FOLDER_TEXT_MOCK = {
     hint: 'A hint text',
     forceIncludeInApi: false,
     iDate: 1717083750000,
-    id: 'b7c41ffd6b6bc1250f2fc85a3637471b',
     indexed: true,
     listed: false,
     modDate: 1717088310000,
@@ -634,6 +633,7 @@ export const HOST_FOLDER_TEXT_MOCK = {
 };
 
 export const CATEGORY_MOCK: DotCMSContentTypeField = {
+    id: '96909fa20a00497cd3b766b52edac0ec',
     clazz: 'com.dotcms.contenttype.model.field.ImmutableWYSIWYGField',
     contentTypeId: '93ebaff75f3e3887bea73ecd04588dc9',
     dataType: 'TEXT',
@@ -643,7 +643,6 @@ export const CATEGORY_MOCK: DotCMSContentTypeField = {
     fixed: false,
     hint: 'A hint text',
     iDate: 1698291913000,
-    id: '96909fa20a00497cd3b766b52edac0ec',
     indexed: false,
     listed: false,
     modDate: 1698291913000,
@@ -658,6 +657,7 @@ export const CATEGORY_MOCK: DotCMSContentTypeField = {
 };
 
 export const CONSTANT_FIELD_MOCK: DotCMSContentTypeField = {
+    id: '666817fcecef5c10cb520c1866baa411',
     clazz: 'com.dotcms.contenttype.model.field.ImmutableConstantField',
     contentTypeId: '93ebaff75f3e3887bea73ecd04588dc9',
     dataType: 'SYSTEM',
@@ -668,7 +668,6 @@ export const CONSTANT_FIELD_MOCK: DotCMSContentTypeField = {
     forceIncludeInApi: false,
     hint: 'this is a hint',
     iDate: 1725491385000,
-    id: '666817fcecef5c10cb520c1866baa411',
     indexed: false,
     listed: false,
     modDate: 1725491385000,
@@ -740,52 +739,52 @@ export const FIELDS_MOCK: DotCMSContentTypeField[] = [
 export const FIELD_MOCK: DotCMSContentTypeField = TEXT_FIELD_MOCK;
 
 export const BINARY_FIELD_CONTENTLET: DotCMSContentlet = {
+    archived: false,
+    baseType: 'CONTENT',
     binaryField:
         '/dA/39de8193694d96c2a6bab783ba9c85b5/binaryField/Screenshot 2023-11-03 at 11.53.40â\u0080¯AM.png',
-    publishDate: '2023-11-07 16:49:24.787',
-    inode: 'd135b73a-8c8f-42ce-bd4e-deb3c067cedd',
-    BinaryContentAsset: '39de8193694d96c2a6bab783ba9c85b5/binaryField',
-    host: '48190c8c-42c4-46af-8d1a-0cd5db894797',
-    locked: false,
-    stInode: 'd1901a41d38b6686dd5ed8f910346d7a',
-    contentType: 'Binary',
-    BinaryMetaData: {
-        modDate: 1699375764242,
-        sha256: '7b4e1c307518ea00e503469e690e4abe42fe1b13aef43cbcbf6eafd9aa532057',
-        length: 136168,
-        title: 'Screenshot 2023-11-03 at 11.53.40â\u0080¯AM.png',
-        version: 20220201,
-        isImage: true,
-        fileSize: 136168,
-        name: 'Screenshot 2023-11-03 at 11.53.40â\u0080¯AM.png',
-        width: 645,
-        contentType: 'image/png',
-        height: 547
-    },
-    identifier: '39de8193694d96c2a6bab783ba9c85b5',
-    folder: 'SYSTEM_FOLDER',
-    hasTitleImage: true,
-    sortOrder: 0,
-    hostName: 'demo.dotcms.com',
-    modDate: '2023-11-07 16:49:24.787',
-    title: '39de8193694d96c2a6bab783ba9c85b5',
-    baseType: 'CONTENT',
-    archived: false,
-    working: true,
-    live: true,
-    owner: 'dotcms.org.1',
-    languageId: 1,
-    url: '/content.d135b73a-8c8f-42ce-bd4e-deb3c067cedd',
-    titleImage: 'binaryField',
-    modUserName: 'Admin User',
-    hasLiveVersion: true,
-    modUser: 'dotcms.org.1',
     binaryFieldVersion:
         '/dA/d135b73a-8c8f-42ce-bd4e-deb3c067cedd/binaryField/Screenshot 2023-11-03 at 11.53.40â\u0080¯AM.png',
-    __icon__: 'contentIcon',
+    BinaryContentAsset: '39de8193694d96c2a6bab783ba9c85b5/binaryField',
+    BinaryMetaData: {
+        contentType: 'image/png',
+        fileSize: 136168,
+        height: 547,
+        isImage: true,
+        length: 136168,
+        modDate: 1699375764242,
+        name: 'Screenshot 2023-11-03 at 11.53.40â\u0080¯AM.png',
+        sha256: '7b4e1c307518ea00e503469e690e4abe42fe1b13aef43cbcbf6eafd9aa532057',
+        title: 'Screenshot 2023-11-03 at 11.53.40â\u0080¯AM.png',
+        version: 20220201,
+        width: 645
+    },
+    contentType: 'Binary',
     contentTypeIcon: 'event_note',
+    folder: 'SYSTEM_FOLDER',
+    hasLiveVersion: true,
+    hasTitleImage: true,
+    host: '48190c8c-42c4-46af-8d1a-0cd5db894797',
+    hostName: 'demo.dotcms.com',
+    __icon__: 'contentIcon',
+    identifier: '39de8193694d96c2a6bab783ba9c85b5',
+    inode: 'd135b73a-8c8f-42ce-bd4e-deb3c067cedd',
+    languageId: 1,
+    live: true,
+    locked: false,
+    modDate: '2023-11-07 16:49:24.787',
+    modUser: 'dotcms.org.1',
+    modUserName: 'Admin User',
+    owner: 'dotcms.org.1',
+    publishDate: '2023-11-07 16:49:24.787',
+    sortOrder: 0,
+    stInode: 'd1901a41d38b6686dd5ed8f910346d7a',
+    title: '39de8193694d96c2a6bab783ba9c85b5',
+    titleImage: 'binaryField',
+    url: '/content.d135b73a-8c8f-42ce-bd4e-deb3c067cedd',
+    value: '/dA/39de8193694d96c2a6bab783ba9c85b5/binaryField/Screenshot 2023-11-03 at 11.53.40â\u0080¯AM.png',
     variant: 'DEFAULT',
-    value: '/dA/39de8193694d96c2a6bab783ba9c85b5/binaryField/Screenshot 2023-11-03 at 11.53.40â\u0080¯AM.png'
+    working: true
 };
 
 /* HELPER FUNCTIONS */
@@ -845,6 +844,7 @@ export const LAYOUT_MOCK: DotCMSContentTypeLayoutRow[] = [
             fieldTypeLabel: 'Row',
             fieldVariables: [],
             fixed: false,
+            forceIncludeInApi: false,
             iDate: 1697051073000,
             id: 'a31ea895f80eb0a3754e4a2292e09a52',
             indexed: false,
@@ -868,6 +868,7 @@ export const LAYOUT_MOCK: DotCMSContentTypeLayoutRow[] = [
                     fieldTypeLabel: 'Column',
                     fieldVariables: [],
                     fixed: false,
+                    forceIncludeInApi: false,
                     iDate: 1697051073000,
                     id: 'd4c32b4b9fb5b11c58c245d4a02bef47',
                     indexed: false,
@@ -891,6 +892,7 @@ export const LAYOUT_MOCK: DotCMSContentTypeLayoutRow[] = [
                         fieldTypeLabel: 'Text',
                         fieldVariables: [],
                         fixed: false,
+                        forceIncludeInApi: false,
                         hint: 'A hint Text',
                         iDate: 1697051093000,
                         id: '1d1505a4569681b923769acb785fd093',
@@ -913,6 +915,7 @@ export const LAYOUT_MOCK: DotCMSContentTypeLayoutRow[] = [
                         fieldTypeLabel: 'Text',
                         fieldVariables: [],
                         fixed: false,
+                        forceIncludeInApi: false,
                         iDate: 1697051107000,
                         id: 'fc776c45044f2d043f5e98eaae36c9ff',
                         indexed: false,
@@ -939,6 +942,7 @@ export const LAYOUT_MOCK: DotCMSContentTypeLayoutRow[] = [
                     fieldTypeLabel: 'Column',
                     fieldVariables: [],
                     fixed: false,
+                    forceIncludeInApi: false,
                     iDate: 1697051077000,
                     id: '848fc78a11e7290efad66eb39333ae2b',
                     indexed: false,
@@ -961,6 +965,7 @@ export const LAYOUT_MOCK: DotCMSContentTypeLayoutRow[] = [
                         fieldTypeLabel: 'Text',
                         fieldVariables: [],
                         fixed: false,
+                        forceIncludeInApi: false,
                         hint: 'A hint text2',
                         iDate: 1697051118000,
                         id: '1f6765de8d4ad069ff308bfca56b9255',
@@ -992,6 +997,7 @@ export const TAB_SINGLE_ROW_MOCK: DotCMSContentTypeLayoutRow = {
         fieldTypeLabel: 'Row',
         fieldVariables: [],
         fixed: false,
+        forceIncludeInApi: false,
         iDate: 1697051073000,
         id: 'a31ea895f80eb0a3754e4a2292e09a52',
         indexed: false,
@@ -1017,6 +1023,7 @@ export const TAB_DIVIDER_MOCK: DotCMSContentTypeLayoutRow = {
         fieldTypeLabel: 'Tab_divider',
         fieldVariables: [],
         fixed: false,
+        forceIncludeInApi: false,
         iDate: 1697051073000,
         id: 'a31ea895f80eb0a3754e4a2292e09a52',
         indexed: false,
@@ -1062,35 +1069,37 @@ export const CONTENT_FORM_DATA_MOCK: DotFormData = {
     } as unknown as DotCMSContentType,
     contentlet: {
         // This contentlet is some random mock, if you need you can change the properties
-        date: MOCK_DATE, // To add the value to the date field, defaultValue is string and I don't think we should change the whole type just for this
-        publishDate: '2023-11-07 16:49:24.787',
-        inode: 'd135b73a-8c8f-42ce-bd4e-deb3c067cedd',
-        host: '48190c8c-42c4-46af-8d1a-0cd5db894797',
-        locked: false,
-        stInode: 'd1901a41d38b6686dd5ed8f910346d7a',
-        contentType: 'Binary',
-        identifier: '39de8193694d96c2a6bab783ba9c85b5',
-        folder: 'SYSTEM_FOLDER',
-        hasTitleImage: true,
-        sortOrder: 0,
-        hostName: 'demo.dotcms.com',
-        modDate: '2023-11-07 16:49:24.787',
-        title: '39de8193694d96c2a6bab783ba9c85b5',
-        baseType: 'CONTENT',
         archived: false,
-        working: true,
-        live: true,
-        owner: 'dotcms.org.1',
+        baseType: 'CONTENT',
+        binaryFieldVersion:
+            '/dA/d135b73a-8c8f-42ce-bd4e-deb3c067cedd/binaryField/Screenshot 2023-11-03 at 11.53.40â\u0080¯AM.png',
+        contentType: 'Binary',
+        contentTypeIcon: 'event_note',
+        date: MOCK_DATE, // To add the value to the date field, defaultValue is string and I don't think we should change the whole type just for this
+        folder: 'SYSTEM_FOLDER',
+        hasLiveVersion: true,
+        hasTitleImage: true,
+        host: '48190c8c-42c4-46af-8d1a-0cd5db894797',
+        hostName: 'demo.dotcms.com',
+        __icon__: 'contentIcon',
+        identifier: '39de8193694d96c2a6bab783ba9c85b5',
+        inode: 'd135b73a-8c8f-42ce-bd4e-deb3c067cedd',
         languageId: 1,
+        live: true,
+        locked: false,
+        modDate: '2023-11-07 16:49:24.787',
+        modUser: 'dotcms.org.1',
+        stInode: 'd1901a41d38b6686dd5ed8f910346d7a',
+        sortOrder: 0,
+        title: '39de8193694d96c2a6bab783ba9c85b5',
+
+        working: true,
+        owner: 'dotcms.org.1',
+
         url: '/content.d135b73a-8c8f-42ce-bd4e-deb3c067cedd',
         titleImage: 'binaryField',
         modUserName: 'Admin User',
-        hasLiveVersion: true,
-        modUser: 'dotcms.org.1',
-        binaryFieldVersion:
-            '/dA/d135b73a-8c8f-42ce-bd4e-deb3c067cedd/binaryField/Screenshot 2023-11-03 at 11.53.40â\u0080¯AM.png',
-        __icon__: 'contentIcon',
-        contentTypeIcon: 'event_note',
+
         variant: 'DEFAULT'
     },
     tabs: []
@@ -1540,1065 +1549,3 @@ export const TEMP_FILE_MOCK: DotCMSTempFile = {
     thumbnailUrl:
         '/contentAsset/image/temp_1e8021f973/tmp/filter/Thumbnail/thumbnail_w/250/thumbnail_h/250/enterprise-angular.pdf'
 };
-
-// ContentType 1 tab, with 1 row and 2 columns
-// 1st column has 1 text field required and 1 text field with hint
-// 2nd column has 1 text field with default value
-// +---------------------------------------------------+
-// |                   Row                             |
-// +-------------------+-------------------------------+
-// |      Column 1     |       Column 2                |
-// +-------------------+-------------------------------+
-// | Text1 (required)  |                               |
-// |                   |                               |
-// | Text2 (with hint) |        Text3                  |
-// |                   |        (with default value)   |
-// +-------------------+-------------------------------+
-
-export const MOCK_CONTENTTYPE_1_TAB: DotCMSContentType = {
-    baseType: 'CONTENT',
-    clazz: 'com.dotcms.contenttype.model.type.ImmutableSimpleContentType',
-    defaultType: false,
-    fields: [
-        {
-            clazz: 'com.dotcms.contenttype.model.field.ImmutableRowField',
-            contentTypeId: '196c8d303e265143806ad19356406ae3',
-            dataType: 'SYSTEM',
-            fieldContentTypeProperties: [],
-            fieldType: 'Row',
-            fieldTypeLabel: 'Row',
-            fieldVariables: [],
-            fixed: false,
-            forceIncludeInApi: false,
-            iDate: 1729012258000,
-            id: '5f992a3eb0d9b92fec736475c4bfb183',
-            indexed: false,
-            listed: false,
-            modDate: 1729012263000,
-            name: 'fields-0',
-            readOnly: false,
-            required: false,
-            searchable: false,
-            sortOrder: 0,
-            unique: false,
-            variable: 'fields0'
-        },
-        {
-            clazz: 'com.dotcms.contenttype.model.field.ImmutableColumnField',
-            contentTypeId: '196c8d303e265143806ad19356406ae3',
-            dataType: 'SYSTEM',
-            fieldContentTypeProperties: [],
-            fieldType: 'Column',
-            fieldTypeLabel: 'Column',
-            fieldVariables: [],
-            fixed: false,
-            forceIncludeInApi: false,
-            iDate: 1729012258000,
-            id: 'b0b0c33b9abf64ecccf51f7900b9af19',
-            indexed: false,
-            listed: false,
-            modDate: 1729012263000,
-            name: 'fields-1',
-            readOnly: false,
-            required: false,
-            searchable: false,
-            sortOrder: 1,
-            unique: false,
-            variable: 'fields1'
-        },
-        {
-            clazz: 'com.dotcms.contenttype.model.field.ImmutableTextField',
-            contentTypeId: '196c8d303e265143806ad19356406ae3',
-            dataType: 'TEXT',
-            fieldType: 'Text',
-            fieldTypeLabel: 'Text',
-            fieldVariables: [],
-            fixed: false,
-            forceIncludeInApi: false,
-            iDate: 1729012281000,
-            id: '69b2ccbb36a0efc135db107eb882d74e',
-            indexed: false,
-            listed: false,
-            modDate: 1729012362000,
-            name: 'text1',
-            readOnly: false,
-            required: true,
-            searchable: false,
-            sortOrder: 2,
-            unique: false,
-            variable: 'text1'
-        },
-        {
-            clazz: 'com.dotcms.contenttype.model.field.ImmutableTextField',
-            contentTypeId: '196c8d303e265143806ad19356406ae3',
-            dataType: 'TEXT',
-            fieldType: 'Text',
-            fieldTypeLabel: 'Text',
-            fieldVariables: [],
-            fixed: false,
-            forceIncludeInApi: false,
-            hint: 'text2 hint',
-            iDate: 1729012287000,
-            id: '4fb628337f5e27ff96ff6ad320d7952b',
-            indexed: false,
-            listed: false,
-            modDate: 1729012373000,
-            name: 'text2',
-            readOnly: false,
-            required: false,
-            searchable: false,
-            sortOrder: 3,
-            unique: false,
-            variable: 'text2'
-        },
-        {
-            clazz: 'com.dotcms.contenttype.model.field.ImmutableColumnField',
-            contentTypeId: '196c8d303e265143806ad19356406ae3',
-            dataType: 'SYSTEM',
-            fieldContentTypeProperties: [],
-            fieldType: 'Column',
-            fieldTypeLabel: 'Column',
-            fieldVariables: [],
-            fixed: false,
-            forceIncludeInApi: false,
-            iDate: 1729012263000,
-            id: '683e570c1fe299628ea10639f354c725',
-            indexed: false,
-            listed: false,
-            modDate: 1729012287000,
-            name: 'fields-2',
-            readOnly: false,
-            required: false,
-            searchable: false,
-            sortOrder: 4,
-            unique: false,
-            variable: 'fields2'
-        },
-        {
-            clazz: 'com.dotcms.contenttype.model.field.ImmutableTextField',
-            contentTypeId: '196c8d303e265143806ad19356406ae3',
-            dataType: 'TEXT',
-            defaultValue: 'default value',
-            fieldType: 'Text',
-            fieldTypeLabel: 'Text',
-            fieldVariables: [],
-            fixed: false,
-            forceIncludeInApi: false,
-            iDate: 1729012357000,
-            id: 'b2d546ae37278b9bb717078be5522a1e',
-            indexed: false,
-            listed: false,
-            modDate: 1729012357000,
-            name: 'text3',
-            readOnly: false,
-            required: false,
-            searchable: false,
-            sortOrder: 5,
-            unique: false,
-            variable: 'text3'
-        }
-    ],
-    fixed: false,
-    folder: 'SYSTEM_FOLDER',
-    host: '48190c8c-42c4-46af-8d1a-0cd5db894797',
-    iDate: 1729012258000,
-    icon: 'event_note',
-    id: '196c8d303e265143806ad19356406ae3',
-    layout: [
-        {
-            divider: {
-                clazz: 'com.dotcms.contenttype.model.field.ImmutableRowField',
-                contentTypeId: '196c8d303e265143806ad19356406ae3',
-                dataType: 'SYSTEM',
-                fieldContentTypeProperties: [],
-                fieldType: 'Row',
-                fieldTypeLabel: 'Row',
-                fieldVariables: [],
-                fixed: false,
-                forceIncludeInApi: false,
-                iDate: 1729012258000,
-                id: '5f992a3eb0d9b92fec736475c4bfb183',
-                indexed: false,
-                listed: false,
-                modDate: 1729012263000,
-                name: 'fields-0',
-                readOnly: false,
-                required: false,
-                searchable: false,
-                sortOrder: 0,
-                unique: false,
-                variable: 'fields0'
-            },
-            columns: [
-                {
-                    columnDivider: {
-                        clazz: 'com.dotcms.contenttype.model.field.ImmutableColumnField',
-                        contentTypeId: '196c8d303e265143806ad19356406ae3',
-                        dataType: 'SYSTEM',
-                        fieldContentTypeProperties: [],
-                        fieldType: 'Column',
-                        fieldTypeLabel: 'Column',
-                        fieldVariables: [],
-                        fixed: false,
-                        forceIncludeInApi: false,
-                        iDate: 1729012258000,
-                        id: 'b0b0c33b9abf64ecccf51f7900b9af19',
-                        indexed: false,
-                        listed: false,
-                        modDate: 1729012263000,
-                        name: 'fields-1',
-                        readOnly: false,
-                        required: false,
-                        searchable: false,
-                        sortOrder: 1,
-                        unique: false,
-                        variable: 'fields1'
-                    },
-                    fields: [
-                        {
-                            clazz: 'com.dotcms.contenttype.model.field.ImmutableTextField',
-                            contentTypeId: '196c8d303e265143806ad19356406ae3',
-                            dataType: 'TEXT',
-                            fieldType: 'Text',
-                            fieldTypeLabel: 'Text',
-                            fieldVariables: [],
-                            fixed: false,
-                            forceIncludeInApi: false,
-                            iDate: 1729012281000,
-                            id: '69b2ccbb36a0efc135db107eb882d74e',
-                            indexed: false,
-                            listed: false,
-                            modDate: 1729012362000,
-                            name: 'text1',
-                            readOnly: false,
-                            required: true,
-                            searchable: false,
-                            sortOrder: 2,
-                            unique: false,
-                            variable: 'text1'
-                        },
-                        {
-                            clazz: 'com.dotcms.contenttype.model.field.ImmutableTextField',
-                            contentTypeId: '196c8d303e265143806ad19356406ae3',
-                            dataType: 'TEXT',
-                            fieldType: 'Text',
-                            fieldTypeLabel: 'Text',
-                            fieldVariables: [],
-                            fixed: false,
-                            forceIncludeInApi: false,
-                            hint: 'text2 hint',
-                            iDate: 1729012287000,
-                            id: '4fb628337f5e27ff96ff6ad320d7952b',
-                            indexed: false,
-                            listed: false,
-                            modDate: 1729012373000,
-                            name: 'text2',
-                            readOnly: false,
-                            required: false,
-                            searchable: false,
-                            sortOrder: 3,
-                            unique: false,
-                            variable: 'text2'
-                        }
-                    ]
-                },
-                {
-                    columnDivider: {
-                        clazz: 'com.dotcms.contenttype.model.field.ImmutableColumnField',
-                        contentTypeId: '196c8d303e265143806ad19356406ae3',
-                        dataType: 'SYSTEM',
-                        fieldContentTypeProperties: [],
-                        fieldType: 'Column',
-                        fieldTypeLabel: 'Column',
-                        fieldVariables: [],
-                        fixed: false,
-                        forceIncludeInApi: false,
-                        iDate: 1729012263000,
-                        id: '683e570c1fe299628ea10639f354c725',
-                        indexed: false,
-                        listed: false,
-                        modDate: 1729012287000,
-                        name: 'fields-2',
-                        readOnly: false,
-                        required: false,
-                        searchable: false,
-                        sortOrder: 4,
-                        unique: false,
-                        variable: 'fields2'
-                    },
-                    fields: [
-                        {
-                            clazz: 'com.dotcms.contenttype.model.field.ImmutableTextField',
-                            contentTypeId: '196c8d303e265143806ad19356406ae3',
-                            dataType: 'TEXT',
-                            fieldType: 'Text',
-                            fieldTypeLabel: 'Text',
-                            fieldVariables: [],
-                            fixed: false,
-                            forceIncludeInApi: false,
-                            iDate: 1729012357000,
-                            id: 'b2d546ae37278b9bb717078be5522a1e',
-                            indexed: false,
-                            listed: false,
-                            modDate: 1729012357000,
-                            name: 'text3',
-                            readOnly: false,
-                            required: false,
-                            searchable: false,
-                            sortOrder: 5,
-                            unique: false,
-                            variable: 'text3'
-                        }
-                    ]
-                }
-            ]
-        }
-    ],
-    metadata: {
-        CONTENT_EDITOR2_ENABLED: false
-    },
-    modDate: 1729012373000,
-    multilingualable: false,
-    name: 'test Mock',
-    system: false,
-    systemActionMappings: {},
-    variable: 'TestMock',
-    versionable: true,
-    workflows: [
-        {
-            archived: false,
-            creationDate: new Date(1729012304196),
-            defaultScheme: false,
-            description: '',
-            entryActionId: null,
-            id: 'd61a59e1-a49c-46f2-a929-db2b4bfa88b2',
-            mandatory: false,
-            modDate: new Date(1728492941334),
-            name: 'System Workflow',
-            system: true
-        }
-    ],
-    nEntries: 0
-};
-
-// Mock of actions of a new ContentType
-export const MOCK_WORKFLOW_ACTIONS_NEW_ITEMNTTYPE_1_TAB: DotCMSWorkflowAction[] = [
-    {
-        assignable: false,
-        commentable: false,
-        condition: '',
-        icon: 'workflowIcon',
-        id: 'ceca71a0-deee-4999-bd47-b01baa1bcfc8',
-        metadata: null,
-        name: 'Save',
-        nextAssign: '654b0931-1027-41f7-ad4d-173115ed8ec1',
-        nextStep: 'ee24a4cb-2d15-4c98-b1bd-6327126451f3',
-        nextStepCurrentStep: false,
-        order: 0,
-        owner: null,
-        roleHierarchyForAssign: false,
-        schemeId: 'd61a59e1-a49c-46f2-a929-db2b4bfa88b2',
-        showOn: ['NEW', 'EDITING', 'LOCKED', 'PUBLISHED', 'UNPUBLISHED'],
-        actionInputs: []
-    },
-    {
-        actionInputs: [],
-        assignable: false,
-        commentable: false,
-        condition: '',
-        icon: 'workflowIcon',
-        id: 'b9d89c80-3d88-4311-8365-187323c96436',
-        metadata: null,
-        name: 'Publish',
-        nextAssign: '654b0931-1027-41f7-ad4d-173115ed8ec1',
-        nextStep: 'dc3c9cd0-8467-404b-bf95-cb7df3fbc293',
-        nextStepCurrentStep: false,
-        order: 0,
-        owner: null,
-        roleHierarchyForAssign: false,
-        schemeId: 'd61a59e1-a49c-46f2-a929-db2b4bfa88b2',
-        showOn: ['LISTING', 'UNLOCKED', 'NEW', 'EDITING', 'LOCKED', 'PUBLISHED', 'UNPUBLISHED']
-    }
-];
-
-// AKA contentlet
-// Contain the values saved from a specific contenttype
-export const MOCK_CONTENTLET_1_TAB: DotCMSContentlet = {
-    archived: false,
-    baseType: 'CONTENT',
-    contentType: 'TestMock',
-    creationDate: '1729016573151',
-    folder: 'SYSTEM_FOLDER',
-    hasLiveVersion: true,
-    hasTitleImage: false,
-    host: '48190c8c-42c4-46af-8d1a-0cd5db894797',
-    hostName: 'demo.dotcms.com',
-    identifier: '2978bb3b66e372b1ffffa9376f33c37b',
-    inode: 'cc120e84-ae80-49d8-9473-36d183d0c1c9',
-    languageId: 1,
-    live: true,
-    locked: false,
-    modDate: '1729024086645',
-    modUser: 'dotcms.org.1',
-    modUserName: 'Admin User',
-    owner: 'dotcms.org.1',
-    ownerName: 'Admin User',
-    publishDate: '1729024086674',
-    publishUser: 'dotcms.org.1',
-    publishUserName: 'Admin User',
-    sortOrder: 0,
-    stInode: '196c8d303e265143806ad19356406ae3',
-    // Start content of the form
-    text1: 'content text 1',
-    text2: 'content text 2',
-    text3: 'default value modified',
-    text11: 'Tab 2 input content', // input in the second tab
-    // end content of the form
-    title: '2978bb3b66e372b1ffffa9376f33c37b',
-    titleImage: 'TITLE_IMAGE_NOT_FOUND',
-    url: '/content.7af6259d-51b0-4c49-9e08-67cbc74100d6',
-    working: true
-};
-
-// ContentType with 2 tabs
-// Tab 1: 1 row and 2 columns
-// 1st column has 1 text field required and 1 text field with hint
-// 2nd column has 1 text field with default value
-// Tab 2: 1 row with 1 column containing 1 text field
-// +---------------------------------------------------+
-// |  Tab 1                |  Tab 2                    |
-// +---------------------------------------------------+
-// |        Row            |        Row                |
-// +----------+------------+---------------------------+
-// | Column 1 | Column 2   |      Column 1             |
-// +----------+------------+---------------------------+
-// | Text1    |            |                           |
-// |(required)|            |                           |
-// |          |   Text3    |       Text1               |
-// | Text2    |  (default  |                           |
-// | (hint)   |   value)   |                           |
-// +----------+------------+---------------------------+
-export const MOCK_CONTENTTYPE_2_TABS: DotCMSContentType = {
-    baseType: 'CONTENT',
-    clazz: 'com.dotcms.contenttype.model.type.ImmutableSimpleContentType',
-    defaultType: false,
-    fields: [
-        {
-            clazz: 'com.dotcms.contenttype.model.field.ImmutableRowField',
-            contentTypeId: '196c8d303e265143806ad19356406ae3',
-            dataType: 'SYSTEM',
-            fieldContentTypeProperties: [],
-            fieldType: 'Row',
-            fieldTypeLabel: 'Row',
-            fieldVariables: [],
-            fixed: false,
-            forceIncludeInApi: false,
-            iDate: 1729012258000,
-            id: '5f992a3eb0d9b92fec736475c4bfb183',
-            indexed: false,
-            listed: false,
-            modDate: 1729012263000,
-            name: 'fields-0',
-            readOnly: false,
-            required: false,
-            searchable: false,
-            sortOrder: 0,
-            unique: false,
-            variable: 'fields0'
-        },
-        {
-            clazz: 'com.dotcms.contenttype.model.field.ImmutableColumnField',
-            contentTypeId: '196c8d303e265143806ad19356406ae3',
-            dataType: 'SYSTEM',
-            fieldContentTypeProperties: [],
-            fieldType: 'Column',
-            fieldTypeLabel: 'Column',
-            fieldVariables: [],
-            fixed: false,
-            forceIncludeInApi: false,
-            iDate: 1729012258000,
-            id: 'b0b0c33b9abf64ecccf51f7900b9af19',
-            indexed: false,
-            listed: false,
-            modDate: 1729012263000,
-            name: 'fields-1',
-            readOnly: false,
-            required: false,
-            searchable: false,
-            sortOrder: 1,
-            unique: false,
-            variable: 'fields1'
-        },
-        {
-            clazz: 'com.dotcms.contenttype.model.field.ImmutableTextField',
-            contentTypeId: '196c8d303e265143806ad19356406ae3',
-            dataType: 'TEXT',
-            fieldType: 'Text',
-            fieldTypeLabel: 'Text',
-            fieldVariables: [],
-            fixed: false,
-            forceIncludeInApi: false,
-            iDate: 1729012281000,
-            id: '69b2ccbb36a0efc135db107eb882d74e',
-            indexed: false,
-            listed: false,
-            modDate: 1729012362000,
-            name: 'text1',
-            readOnly: false,
-            required: true,
-            searchable: false,
-            sortOrder: 2,
-            unique: false,
-            variable: 'text1'
-        },
-        {
-            clazz: 'com.dotcms.contenttype.model.field.ImmutableTextField',
-            contentTypeId: '196c8d303e265143806ad19356406ae3',
-            dataType: 'TEXT',
-            fieldType: 'Text',
-            fieldTypeLabel: 'Text',
-            fieldVariables: [],
-            fixed: false,
-            forceIncludeInApi: false,
-            hint: 'text2 hint',
-            iDate: 1729012287000,
-            id: '4fb628337f5e27ff96ff6ad320d7952b',
-            indexed: false,
-            listed: false,
-            modDate: 1729012373000,
-            name: 'text2',
-            readOnly: false,
-            required: false,
-            searchable: false,
-            sortOrder: 3,
-            unique: false,
-            variable: 'text2'
-        },
-        {
-            clazz: 'com.dotcms.contenttype.model.field.ImmutableColumnField',
-            contentTypeId: '196c8d303e265143806ad19356406ae3',
-            dataType: 'SYSTEM',
-            fieldContentTypeProperties: [],
-            fieldType: 'Column',
-            fieldTypeLabel: 'Column',
-            fieldVariables: [],
-            fixed: false,
-            forceIncludeInApi: false,
-            iDate: 1729012263000,
-            id: '683e570c1fe299628ea10639f354c725',
-            indexed: false,
-            listed: false,
-            modDate: 1729012287000,
-            name: 'fields-2',
-            readOnly: false,
-            required: false,
-            searchable: false,
-            sortOrder: 4,
-            unique: false,
-            variable: 'fields2'
-        },
-        {
-            clazz: 'com.dotcms.contenttype.model.field.ImmutableTextField',
-            contentTypeId: '196c8d303e265143806ad19356406ae3',
-            dataType: 'TEXT',
-            defaultValue: 'default value',
-            fieldType: 'Text',
-            fieldTypeLabel: 'Text',
-            fieldVariables: [],
-            fixed: false,
-            forceIncludeInApi: false,
-            iDate: 1729012357000,
-            id: 'b2d546ae37278b9bb717078be5522a1e',
-            indexed: false,
-            listed: false,
-            modDate: 1729023673000,
-            name: 'text3',
-            readOnly: false,
-            required: false,
-            searchable: false,
-            sortOrder: 5,
-            unique: false,
-            variable: 'text3'
-        },
-        {
-            clazz: 'com.dotcms.contenttype.model.field.ImmutableTabDividerField',
-            contentTypeId: '196c8d303e265143806ad19356406ae3',
-            dataType: 'SYSTEM',
-            fieldType: 'Tab_divider',
-            fieldTypeLabel: 'Tab Divider',
-            fieldVariables: [],
-            fixed: false,
-            forceIncludeInApi: false,
-            iDate: 1729025952000,
-            id: '263d6fa59a75fe9fc33ca929903c34fe',
-            indexed: false,
-            listed: false,
-            modDate: 1729025952000,
-            name: 'New Tab',
-            readOnly: false,
-            required: false,
-            searchable: false,
-            sortOrder: 6,
-            unique: false,
-            variable: 'newTab'
-        },
-        {
-            clazz: 'com.dotcms.contenttype.model.field.ImmutableRowField',
-            contentTypeId: '196c8d303e265143806ad19356406ae3',
-            dataType: 'SYSTEM',
-            fieldContentTypeProperties: [],
-            fieldType: 'Row',
-            fieldTypeLabel: 'Row',
-            fieldVariables: [],
-            fixed: false,
-            forceIncludeInApi: false,
-            iDate: 1729025973000,
-            id: 'b699c53c74fedd58e866627a09a1f3b7',
-            indexed: false,
-            listed: false,
-            modDate: 1729025973000,
-            name: 'fields-3',
-            readOnly: false,
-            required: false,
-            searchable: false,
-            sortOrder: 7,
-            unique: false,
-            variable: 'fields3'
-        },
-        {
-            clazz: 'com.dotcms.contenttype.model.field.ImmutableColumnField',
-            contentTypeId: '196c8d303e265143806ad19356406ae3',
-            dataType: 'SYSTEM',
-            fieldContentTypeProperties: [],
-            fieldType: 'Column',
-            fieldTypeLabel: 'Column',
-            fieldVariables: [],
-            fixed: false,
-            forceIncludeInApi: false,
-            iDate: 1729025973000,
-            id: '76c04624b88dbb58207a4a40bbac4bca',
-            indexed: false,
-            listed: false,
-            modDate: 1729025973000,
-            name: 'fields-4',
-            readOnly: false,
-            required: false,
-            searchable: false,
-            sortOrder: 8,
-            unique: false,
-            variable: 'fields4'
-        },
-        {
-            clazz: 'com.dotcms.contenttype.model.field.ImmutableTextField',
-            contentTypeId: '196c8d303e265143806ad19356406ae3',
-            dataType: 'TEXT',
-            fieldType: 'Text',
-            fieldTypeLabel: 'Text',
-            fieldVariables: [],
-            fixed: false,
-            forceIncludeInApi: false,
-            iDate: 1729025973000,
-            id: '883619ed3cea50a1ed4cdd76366a4b3c',
-            indexed: false,
-            listed: false,
-            modDate: 1729025973000,
-            name: 'Text 1',
-            readOnly: false,
-            required: false,
-            searchable: false,
-            sortOrder: 9,
-            unique: false,
-            variable: 'text11'
-        }
-    ],
-    fixed: false,
-    folder: 'SYSTEM_FOLDER',
-
-    host: '48190c8c-42c4-46af-8d1a-0cd5db894797',
-    iDate: 1729012258000,
-    icon: 'event_note',
-    id: '196c8d303e265143806ad19356406ae3',
-    layout: [
-        {
-            divider: {
-                clazz: 'com.dotcms.contenttype.model.field.ImmutableRowField',
-                contentTypeId: '196c8d303e265143806ad19356406ae3',
-                dataType: 'SYSTEM',
-                fieldContentTypeProperties: [],
-                fieldType: 'Row',
-                fieldTypeLabel: 'Row',
-                fieldVariables: [],
-                fixed: false,
-                forceIncludeInApi: false,
-                iDate: 1729012258000,
-                id: '5f992a3eb0d9b92fec736475c4bfb183',
-                indexed: false,
-                listed: false,
-                modDate: 1729012263000,
-                name: 'fields-0',
-                readOnly: false,
-                required: false,
-                searchable: false,
-                sortOrder: 0,
-                unique: false,
-                variable: 'fields0'
-            },
-            columns: [
-                {
-                    columnDivider: {
-                        clazz: 'com.dotcms.contenttype.model.field.ImmutableColumnField',
-                        contentTypeId: '196c8d303e265143806ad19356406ae3',
-                        dataType: 'SYSTEM',
-                        fieldContentTypeProperties: [],
-                        fieldType: 'Column',
-                        fieldTypeLabel: 'Column',
-                        fieldVariables: [],
-                        fixed: false,
-                        forceIncludeInApi: false,
-                        iDate: 1729012258000,
-                        id: 'b0b0c33b9abf64ecccf51f7900b9af19',
-                        indexed: false,
-                        listed: false,
-                        modDate: 1729012263000,
-                        name: 'fields-1',
-                        readOnly: false,
-                        required: false,
-                        searchable: false,
-                        sortOrder: 1,
-                        unique: false,
-                        variable: 'fields1'
-                    },
-                    fields: [
-                        {
-                            clazz: 'com.dotcms.contenttype.model.field.ImmutableTextField',
-                            contentTypeId: '196c8d303e265143806ad19356406ae3',
-                            dataType: 'TEXT',
-                            fieldType: 'Text',
-                            fieldTypeLabel: 'Text',
-                            fieldVariables: [],
-                            fixed: false,
-                            forceIncludeInApi: false,
-                            iDate: 1729012281000,
-                            id: '69b2ccbb36a0efc135db107eb882d74e',
-                            indexed: false,
-                            listed: false,
-                            modDate: 1729012362000,
-                            name: 'text1',
-                            readOnly: false,
-                            required: true,
-                            searchable: false,
-                            sortOrder: 2,
-                            unique: false,
-                            variable: 'text1'
-                        },
-                        {
-                            clazz: 'com.dotcms.contenttype.model.field.ImmutableTextField',
-                            contentTypeId: '196c8d303e265143806ad19356406ae3',
-                            dataType: 'TEXT',
-                            fieldType: 'Text',
-                            fieldTypeLabel: 'Text',
-                            fieldVariables: [],
-                            fixed: false,
-                            forceIncludeInApi: false,
-                            hint: 'text2 hint',
-                            iDate: 1729012287000,
-                            id: '4fb628337f5e27ff96ff6ad320d7952b',
-                            indexed: false,
-                            listed: false,
-                            modDate: 1729012373000,
-                            name: 'text2',
-                            readOnly: false,
-                            required: false,
-                            searchable: false,
-                            sortOrder: 3,
-                            unique: false,
-                            variable: 'text2'
-                        }
-                    ]
-                },
-                {
-                    columnDivider: {
-                        clazz: 'com.dotcms.contenttype.model.field.ImmutableColumnField',
-                        contentTypeId: '196c8d303e265143806ad19356406ae3',
-                        dataType: 'SYSTEM',
-                        fieldContentTypeProperties: [],
-                        fieldType: 'Column',
-                        fieldTypeLabel: 'Column',
-                        fieldVariables: [],
-                        fixed: false,
-                        forceIncludeInApi: false,
-                        iDate: 1729012263000,
-                        id: '683e570c1fe299628ea10639f354c725',
-                        indexed: false,
-                        listed: false,
-                        modDate: 1729012287000,
-                        name: 'fields-2',
-                        readOnly: false,
-                        required: false,
-                        searchable: false,
-                        sortOrder: 4,
-                        unique: false,
-                        variable: 'fields2'
-                    },
-                    fields: [
-                        {
-                            clazz: 'com.dotcms.contenttype.model.field.ImmutableTextField',
-                            contentTypeId: '196c8d303e265143806ad19356406ae3',
-                            dataType: 'TEXT',
-                            defaultValue: 'default value',
-                            fieldType: 'Text',
-                            fieldTypeLabel: 'Text',
-                            fieldVariables: [],
-                            fixed: false,
-                            forceIncludeInApi: false,
-                            iDate: 1729012357000,
-                            id: 'b2d546ae37278b9bb717078be5522a1e',
-                            indexed: false,
-                            listed: false,
-                            modDate: 1729023673000,
-                            name: 'text3',
-                            readOnly: false,
-                            required: false,
-                            searchable: false,
-                            sortOrder: 5,
-                            unique: false,
-                            variable: 'text3'
-                        }
-                    ]
-                }
-            ]
-        },
-        {
-            divider: {
-                clazz: 'com.dotcms.contenttype.model.field.ImmutableTabDividerField',
-                contentTypeId: '196c8d303e265143806ad19356406ae3',
-                dataType: 'SYSTEM',
-                fieldType: 'Tab_divider',
-                fieldTypeLabel: 'Tab Divider',
-                fieldVariables: [],
-                fixed: false,
-                forceIncludeInApi: false,
-                iDate: 1729025952000,
-                id: '263d6fa59a75fe9fc33ca929903c34fe',
-                indexed: false,
-                listed: false,
-                modDate: 1729025952000,
-                name: 'New Tab',
-                readOnly: false,
-                required: false,
-                searchable: false,
-                sortOrder: 6,
-                unique: false,
-                variable: 'newTab'
-            },
-            columns: []
-        },
-        {
-            divider: {
-                clazz: 'com.dotcms.contenttype.model.field.ImmutableRowField',
-                contentTypeId: '196c8d303e265143806ad19356406ae3',
-                dataType: 'SYSTEM',
-                fieldContentTypeProperties: [],
-                fieldType: 'Row',
-                fieldTypeLabel: 'Row',
-                fieldVariables: [],
-                fixed: false,
-                forceIncludeInApi: false,
-                iDate: 1729025973000,
-                id: 'b699c53c74fedd58e866627a09a1f3b7',
-                indexed: false,
-                listed: false,
-                modDate: 1729025973000,
-                name: 'fields-3',
-                readOnly: false,
-                required: false,
-                searchable: false,
-                sortOrder: 7,
-                unique: false,
-                variable: 'fields3'
-            },
-            columns: [
-                {
-                    columnDivider: {
-                        clazz: 'com.dotcms.contenttype.model.field.ImmutableColumnField',
-                        contentTypeId: '196c8d303e265143806ad19356406ae3',
-                        dataType: 'SYSTEM',
-                        fieldContentTypeProperties: [],
-                        fieldType: 'Column',
-                        fieldTypeLabel: 'Column',
-                        fieldVariables: [],
-                        fixed: false,
-                        forceIncludeInApi: false,
-                        iDate: 1729025973000,
-                        id: '76c04624b88dbb58207a4a40bbac4bca',
-                        indexed: false,
-                        listed: false,
-                        modDate: 1729025973000,
-                        name: 'fields-4',
-                        readOnly: false,
-                        required: false,
-                        searchable: false,
-                        sortOrder: 8,
-                        unique: false,
-                        variable: 'fields4'
-                    },
-                    fields: [
-                        {
-                            clazz: 'com.dotcms.contenttype.model.field.ImmutableTextField',
-                            contentTypeId: '196c8d303e265143806ad19356406ae3',
-                            dataType: 'TEXT',
-                            fieldType: 'Text',
-                            fieldTypeLabel: 'Text',
-                            fieldVariables: [],
-                            fixed: false,
-                            forceIncludeInApi: false,
-                            iDate: 1729025973000,
-                            id: '883619ed3cea50a1ed4cdd76366a4b3c',
-                            indexed: false,
-                            listed: false,
-                            modDate: 1729025973000,
-                            name: 'Text 1',
-                            readOnly: false,
-                            required: false,
-                            searchable: false,
-                            sortOrder: 9,
-                            unique: false,
-                            variable: 'text11'
-                        }
-                    ]
-                }
-            ]
-        }
-    ],
-    metadata: {
-        CONTENT_EDITOR2_ENABLED: true
-    },
-    modDate: 1729025973000,
-    multilingualable: false,
-    name: 'test Mock',
-    system: false,
-    systemActionMappings: {},
-    variable: 'TestMock',
-    versionable: true,
-    workflows: [
-        {
-            archived: false,
-            creationDate: new Date(1729012842268),
-            defaultScheme: false,
-            description: '',
-            entryActionId: null,
-            id: 'd61a59e1-a49c-46f2-a929-db2b4bfa88b2',
-            mandatory: false,
-            modDate: new Date(1728492941334),
-            name: 'System Workflow',
-            system: true
-        }
-    ],
-    nEntries: 0
-};
-
-/**
- * Fields allowed to be used in form controls from the MOCK_CONTENTTYPE_2_TABS Content Type.
- * 
- * This array contains a subset of fields from MOCK_CONTENTTYPE_2_TABS that are suitable
- * for form controls. It excludes fields of types defined in NON_FORM_CONTROL_FIELD_TYPES,
- * such as Row, Column, Tab_divider, Constant-Field, and Hidden-Field.
-
- * 
- * @see MOCK_CONTENTTYPE_2_TABS
- * @see NON_FORM_CONTROL_FIELD_TYPES
- */
-export const MOCK_FORM_CONTROL_FIELDS: DotCMSContentTypeField[] = [
-    {
-        clazz: 'com.dotcms.contenttype.model.field.ImmutableTextField',
-        contentTypeId: '196c8d303e265143806ad19356406ae3',
-        dataType: 'TEXT',
-        fieldType: 'Text',
-        fieldTypeLabel: 'Text',
-        fieldVariables: [],
-        fixed: false,
-        forceIncludeInApi: false,
-        iDate: 1729012281000,
-        id: '69b2ccbb36a0efc135db107eb882d74e',
-        indexed: false,
-        listed: false,
-        modDate: 1729012362000,
-        name: 'text1',
-        readOnly: false,
-        required: true,
-        searchable: false,
-        sortOrder: 2,
-        unique: false,
-        variable: 'text1'
-    },
-    {
-        clazz: 'com.dotcms.contenttype.model.field.ImmutableTextField',
-        contentTypeId: '196c8d303e265143806ad19356406ae3',
-        dataType: 'TEXT',
-        fieldType: 'Text',
-        fieldTypeLabel: 'Text',
-        fieldVariables: [],
-        fixed: false,
-        forceIncludeInApi: false,
-        hint: 'text2 hint',
-        iDate: 1729012287000,
-        id: '4fb628337f5e27ff96ff6ad320d7952b',
-        indexed: false,
-        listed: false,
-        modDate: 1729012373000,
-        name: 'text2',
-        readOnly: false,
-        required: false,
-        searchable: false,
-        sortOrder: 3,
-        unique: false,
-        variable: 'text2'
-    },
-
-    {
-        clazz: 'com.dotcms.contenttype.model.field.ImmutableTextField',
-        contentTypeId: '196c8d303e265143806ad19356406ae3',
-        dataType: 'TEXT',
-        defaultValue: 'default value',
-        fieldType: 'Text',
-        fieldTypeLabel: 'Text',
-        fieldVariables: [],
-        fixed: false,
-        forceIncludeInApi: false,
-        iDate: 1729012357000,
-        id: 'b2d546ae37278b9bb717078be5522a1e',
-        indexed: false,
-        listed: false,
-        modDate: 1729023673000,
-        name: 'text3',
-        readOnly: false,
-        required: false,
-        searchable: false,
-        sortOrder: 5,
-        unique: false,
-        variable: 'text3'
-    },
-
-    {
-        clazz: 'com.dotcms.contenttype.model.field.ImmutableTextField',
-        contentTypeId: '196c8d303e265143806ad19356406ae3',
-        dataType: 'TEXT',
-        fieldType: 'Text',
-        fieldTypeLabel: 'Text',
-        fieldVariables: [],
-        fixed: false,
-        forceIncludeInApi: false,
-        iDate: 1729025973000,
-        id: '883619ed3cea50a1ed4cdd76366a4b3c',
-        indexed: false,
-        listed: false,
-        modDate: 1729025973000,
-        name: 'Text 1',
-        readOnly: false,
-        required: false,
-        searchable: false,
-        sortOrder: 9,
-        unique: false,
-        variable: 'text11'
-    }
-];
